+++ src/Farmer/Common.fs
@@ -448,7 +448,7 @@ module WebApp =
         let Logging = ExtensionName "Microsoft.AspNetCore.AzureAppServices.SiteExtension"
 
 module CognitiveServices =
-    /// Type of SKU. See https://github.com/Azure/azure-quickstart-templates/tree/master/101-cognitive-services-translate
+    /// Type of SKU. See https://docs.microsoft.com/en-us/rest/api/cognitiveservices/accountmanagement/resourceskus/list
     type Sku =
         /// Free Tier
         | F0
@@ -461,7 +461,6 @@ module CognitiveServices =
     type Kind =
         | AllInOne
         | AnomalyDetector
-        | Bing_Autosuggest_v7 | Bing_CustomSearch | Bing_EntitySearch | Bing_Search_v7 | Bing_SpellCheck_v7
         | CognitiveServices
         | ComputerVision
         | ContentModerator
@@ -477,6 +476,22 @@ module CognitiveServices =
         | SpeechServices
         | TextAnalytics
         | TextTranslation
+
+module BingSearch =
+    /// Type of SKU. See https://www.microsoft.com/en-us/bing/apis/pricing
+    type Sku =
+        /// Free Tier
+        | F1
+        | S0
+        | S1
+        | S2
+        | S3
+        | S4
+        | S5
+        | S6
+        | S7
+        | S8
+        | S9
 
 module ContainerRegistry =
     /// Container Registry SKU